refactor/optimize(guis): Moved commonly used methods to a GuiHelper object & have variables save to ram instead of constantly getting them.
